requirements.txt
docker-py==1.5.0
gevent==1.0.2
urwid==1.3.0
